refactor(web): make task status in the meta panel display-only
Show the task status as a read-only color-coded TaskStatusBadge in the
meta panel instead of an editable dropdown. Status changes still flow
through the existing Close / Re-open controls; the panel just surfaces the
current state.
